+++ modules/compute/eks/stack.py
@@ -31,6 +31,7 @@ from helpers import (
     get_chart_version,
     get_image,
 )
+from utils.k8s import convert_node_labels_to_k8sargs, convert_taints_to_k8sargs
 
 project_dir = os.path.dirname(os.path.abspath(__file__))
 
@@ -342,6 +343,17 @@ class Eks(Stack):  # type: ignore
         """
         Creates a Self Managed Node Group with the specified configuration.
         """
+
+        KUBELET_EXTRA_ARGS = ""
+
+        if ng_config.get("eks_node_labels"):
+            labels = convert_node_labels_to_k8sargs(ng_config.get("eks_node_labels"))
+            KUBELET_EXTRA_ARGS = f"--node-labels {labels}"
+
+        if ng_config.get("eks_node_taints"):
+            taints = convert_taints_to_k8sargs(ng_config.get("eks_node_taints"))
+            KUBELET_EXTRA_ARGS = f"{KUBELET_EXTRA_ARGS} --register-with-taints {taints}"
+
         self_managed_nodegroup = eks_cluster.add_auto_scaling_group_capacity(
             f"self-managed-{ng_config.get('eks_ng_name')}",
             desired_capacity=ng_config.get("eks_node_quantity"),
@@ -349,6 +361,9 @@ class Eks(Stack):  # type: ignore
             min_capacity=ng_config.get("eks_node_min_quantity"),
             update_policy=None,
             instance_type=ec2.InstanceType(str(ng_config.get("eks_node_instance_type"))),
+            bootstrap_options=eks.BootstrapOptions(
+                kubelet_extra_args=KUBELET_EXTRA_ARGS if KUBELET_EXTRA_ARGS else None,
+            ),
             vpc_subnets=ec2.SubnetSelection(subnets=self.dataplane_subnets),
             auto_scaling_group_name=f"{dep_mod}-{ng_config.get('eks_ng_name')}",
             group_metrics=[asg.GroupMetrics.all()],
@@ -461,6 +476,7 @@ class Eks(Stack):  # type: ignore
         """
         Creates an Amazon EKS cluster with the specified configuration.
         """
+
         # Create the EKS cluster
         eks_cluster = eks.Cluster(
             self,