+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - emr-on-eks
   - buckets
 - added support for `nvidia device plugin` for management of GPUs on EKS clusters
+- add integration tests for module `storage/efs`
 
 ### **Changed**