+++ CHANGELOG.md
@@ -17,11 +17,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - emr-on-eks
   - buckets
 - added support for `nvidia device plugin` for management of GPUs on EKS clusters
+- added support for adding taints and labels for self managed ngs
 
 ### **Changed**
 
 - refactored the deployspec of `emr-on-eks` module and used seedfarmer to add/group metadata of 2 stacks
 - fixed the readme and app.py of `emr-serverless` module
+- fixed the missing taints and labels for self managed nodegroups of `eks` module
 
 ### **Removed**