+++ CHANGELOG.md
@@ -18,7 +18,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - buckets
 - added support for `nvidia device plugin` for management of GPUs on EKS clusters
 - added support for adding taints and labels for self managed ngs
-- added support for whitelisting CIDRs on EKS API Endpoint
+- added a custom timeout for Self Managed Nodegroups for successful signalling
+- configured custom allocation strategies for AWS Batch module
 
 ### **Changed**