+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 =======
 
-## UNRELEASED
+## v1.9.0 (2024-07-10)
 
 ### **Added**